Tried out mixed precision matrix
In theory, this would be ideal, because 1e38 is enough scale, just 1e7 not enough precision for positions

+++ Bullet/src/me/anno/bullet/BulletDebugDraw.kt
@@ -15,7 +15,7 @@ object BulletDebugDraw : IDebugDraw() {
 
     var stack = Matrix4f()
     var cam = org.joml.Vector3d()
-    val worldScale: Double get() = RenderState.worldScale
+    val worldScale: Float get() = RenderState.worldScale
 
     /**
     public static final int NO_DEBUG              = 0;
